medusa/requirements.txt
kafka-python==2.0.2
networkx==2.6.3
